requirements.txt
Flask==2.2.5
Jinja2==3.1.2
httpx==0.24.1
SQLAlchemy==1.4.49
psycopg2-binary==2.9.7
alembic==1.11.1
redis==4.5.5
Flask-Limiter==2.8.1
prometheus-client==0.16.0
structlog==23.3.0
tenacity==8.2.2
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.22.0
gunicorn==20.1.0
opentelemetry-api==1.24.0
opentelemetry-instrumentation-flask==0.37b0
requests
